2023 WIP rework

+++ src/measures.cpp
@@ -97,7 +97,7 @@ bool update_measures(void *)
 		measures.water_temperature_raw = water_get_temperature();
 
 		// Make real measure only if pump is ON for on time > FILTER_PUMP_ON_MIN_TIME_S
-		if ((pump_filtration_is_on()) && ((abs(millis() - state.filter_time_pump_on) / 1000) >= FILTER_PUMP_ON_MIN_TIME_S))
+		if ((pump_filtration_is_on()) && ((abs((int)(millis() - state.filter_time_pump_on)) / 1000) >= FILTER_PUMP_ON_MIN_TIME_S))
 		{
 			measures.water_temperature = measures.water_temperature_raw;
 			if (measures.water_temperature_raw > measures.day_max_water_temperature)
